Added alt, width, height support for img tag, fixes
Added alt, width, and height parameter support for the img tag.

Also added a fix to properly count the number of parameter matches in
both the image and quote tag. Originally it wasn't working and would
allow tags with multiple of the same parameter though which caused some
parsing issues.