Removed the date fix in quote
Removed the previous commit's date fix for quotes because would may use
the quote tag in the future and the date being passed in will be in
milliseconds. This would be inaccurately parsed.

The fix is to have everything in millisecond scale. Thus the import
functions in epochtalk will convert all SMF times to millisecond scale
so that the parser will always work with JS's Date object without
modification.